pyproject.toml
[tool.poetry]
name = "giz"
version = "0.1.0"
description = ""
authors = ["Thales Macedo Garitezi <thalesmg@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyGithub = "^1.54.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
giz = "giz.__main__:main"
